refactor some dialog actions
